third_party/WebKit/LayoutTests/external/wpt/css/geometry/DOMMatrix2DInit-validate-fixup-expected.txt

This is a testharness.js-based test.
Found 184 tests; 93 PASS, 91 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS setTransform({a: 1, m11: 2}) (invalid)
PASS addPath({a: 1, m11: 2}) (invalid)
PASS setTransform({b: 0, m12: -1}) (invalid)
PASS addPath({b: 0, m12: -1}) (invalid)
PASS setTransform({c: Infinity, m21: -Infinity}) (invalid)
PASS addPath({c: Infinity, m21: -Infinity}) (invalid)
PASS setTransform({d: 0, m22: NaN}) (invalid)
PASS addPath({d: 0, m22: NaN}) (invalid)
PASS setTransform({e: 1, m41: 1.00000001}) (invalid)
PASS addPath({e: 1, m41: 1.00000001}) (invalid)
PASS setTransform({f: 0, m42: 5e-324}) (invalid)
PASS addPath({f: 0, m42: 5e-324}) (invalid)
PASS setTransform (Sanity check without dictionary)
PASS addPath (Sanity check without second parameter)
PASS setTransform({m13: 1, is2D: true})
FAIL addPath({m13: 1, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m14: 1, is2D: true})
FAIL addPath({m14: 1, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m23: 1, is2D: true})
FAIL addPath({m23: 1, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m24: 1, is2D: true})
FAIL addPath({m24: 1, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m31: 1, is2D: true})
FAIL addPath({m31: 1, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m32: 1, is2D: true})
FAIL addPath({m32: 1, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m33: 0, is2D: true})
FAIL addPath({m33: 0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m33: -0, is2D: true})
FAIL addPath({m33: -0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m33: -1, is2D: true})
FAIL addPath({m33: -1, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m34: 1, is2D: true})
FAIL addPath({m34: 1, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m43: 1, is2D: true})
FAIL addPath({m43: 1, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m44: 0, is2D: true})
FAIL addPath({m44: 0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({})
FAIL addPath({}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({is2D: undefined})
FAIL addPath({is2D: undefined}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({a: 1, m11: 1})
FAIL addPath({a: 1, m11: 1}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({b: 0, m12: undefined})
FAIL addPath({b: 0, m12: undefined}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({c: 0, m21: 0})
FAIL addPath({c: 0, m21: 0}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
FAIL setTransform({c: 0, m21: -0}) assert_equals: m21 expected -0 but got 0
FAIL addPath({c: 0, m21: -0}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({c: -0, m21: 0})
FAIL addPath({c: -0, m21: 0}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
FAIL setTransform({c: -0, m21: -0}) assert_equals: m21 expected -0 but got 0
FAIL addPath({c: -0, m21: -0}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({d: Infinity, m22: Infinity})
FAIL addPath({d: Infinity, m22: Infinity}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({e: -Infinity, m41: -Infinity})
FAIL addPath({e: -Infinity, m41: -Infinity}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({f: NaN, m42: NaN})
FAIL addPath({f: NaN, m42: NaN}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({f: NaN, m42: NaN, is2D: true})
FAIL addPath({f: NaN, m42: NaN, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({f: 0, m42: null})
FAIL addPath({f: 0, m42: null}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({f: -0, m42: null})
FAIL addPath({f: -0, m42: null}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({a: 2})
FAIL addPath({a: 2}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({b: 2})
FAIL addPath({b: 2}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({c: 2})
FAIL addPath({c: 2}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({d: 2})
FAIL addPath({d: 2}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({e: 2})
FAIL addPath({e: 2}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({f: 2})
FAIL addPath({f: 2}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
FAIL setTransform({a: -0, b: -0, c: -0, d: -0, e: -0, f: -0}) assert_equals: m41 expected -0 but got 0
FAIL addPath({a: -0, b: -0, c: -0, d: -0, e: -0, f: -0}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
FAIL setTransform({a: -0, b: -0, c: -0, d: -0, e: -0, f: -0, is2D: true}) assert_equals: m41 expected -0 but got 0
FAIL addPath({a: -0, b: -0, c: -0, d: -0, e: -0, f: -0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m11: 2})
FAIL addPath({m11: 2}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m12: 2})
FAIL addPath({m12: 2}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m21: 2})
FAIL addPath({m21: 2}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m22: 2})
FAIL addPath({m22: 2}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m41: 2})
FAIL addPath({m41: 2}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m42: 2})
FAIL addPath({m42: 2}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
FAIL setTransform({m11: -0, m12: -0, m21: -0, m22: -0, m41: -0, m42: -0}) assert_equals: m41 expected -0 but got 0
FAIL addPath({m11: -0, m12: -0, m21: -0, m22: -0, m41: -0, m42: -0}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
FAIL setTransform({m11: -0, m12: -0, m21: -0, m22: -0, m41: -0, m42: -0, is2D: true}) assert_equals: m41 expected -0 but got 0
FAIL addPath({m11: -0, m12: -0, m21: -0, m22: -0, m41: -0, m42: -0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m13: 0, is2D: true})
FAIL addPath({m13: 0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m13: -0, is2D: true})
FAIL addPath({m13: -0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m14: 0, is2D: true})
FAIL addPath({m14: 0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m14: -0, is2D: true})
FAIL addPath({m14: -0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m23: 0, is2D: true})
FAIL addPath({m23: 0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m23: -0, is2D: true})
FAIL addPath({m23: -0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m24: 0, is2D: true})
FAIL addPath({m24: 0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m24: -0, is2D: true})
FAIL addPath({m24: -0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m31: 0, is2D: true})
FAIL addPath({m31: 0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m31: -0, is2D: true})
FAIL addPath({m31: -0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m32: 0, is2D: true})
FAIL addPath({m32: 0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m32: -0, is2D: true})
FAIL addPath({m32: -0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m33: 1, is2D: true})
FAIL addPath({m33: 1, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m34: 0, is2D: true})
FAIL addPath({m34: 0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m34: -0, is2D: true})
FAIL addPath({m34: -0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m43: 0, is2D: true})
FAIL addPath({m43: 0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m43: -0, is2D: true})
FAIL addPath({m43: -0, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m44: 1, is2D: true})
FAIL addPath({m44: 1, is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({is2D: true})
FAIL addPath({is2D: true}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m13: 1, is2D: false})
FAIL addPath({m13: 1, is2D: false}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m14: 1, is2D: false})
FAIL addPath({m14: 1, is2D: false}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m23: 1, is2D: false})
FAIL addPath({m23: 1, is2D: false}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m24: 1, is2D: false})
FAIL addPath({m24: 1, is2D: false}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m31: 1, is2D: false})
FAIL addPath({m31: 1, is2D: false}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m32: 1, is2D: false})
FAIL addPath({m32: 1, is2D: false}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m33: 0, is2D: false})
FAIL addPath({m33: 0, is2D: false}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m33: -0, is2D: false})
FAIL addPath({m33: -0, is2D: false}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m33: -1, is2D: false})
FAIL addPath({m33: -1, is2D: false}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m34: 1, is2D: false})
FAIL addPath({m34: 1, is2D: false}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m43: 1, is2D: false})
FAIL addPath({m43: 1, is2D: false}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m44: 0, is2D: false})
FAIL addPath({m44: 0, is2D: false}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m13: 1})
FAIL addPath({m13: 1}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m14: 1})
FAIL addPath({m14: 1}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m23: 1})
FAIL addPath({m23: 1}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m24: 1})
FAIL addPath({m24: 1}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m31: 1})
FAIL addPath({m31: 1}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m32: 1})
FAIL addPath({m32: 1}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m33: 0})
FAIL addPath({m33: 0}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m34: 1})
FAIL addPath({m34: 1}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m43: 1})
FAIL addPath({m43: 1}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({m44: 0})
FAIL addPath({m44: 0}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({is2D: false})
FAIL addPath({is2D: false}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
PASS setTransform({is2D: null})
FAIL addPath({is2D: null}) Failed to execute 'addPath' on 'Path2D': parameter 2 is not of type 'SVGMatrix'.
Harness: the test ran to completion.
